WifiHal: Bring up virtual interface created for STA mode after creation
wifi_virtual_interface_create() is used to create additional interface
for secondary STA instance in case of STA+STA concurrency.

Bring up the interface after creating it so that the driver accepts QCA
vendor commands such as set_link_layer_stats which needs the netdev to
be in IFF_RUNNING state.

Change-Id: I5ad5b5ee372a6311836c5f12d55566a200079793
CRs-Fixed: 3055721